requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
fake-useragent>=1.4.0
lxml>=4.9.0
trafilatura>=1.6.1
readability-lxml>=0.8.1
PyPDF2>=3.0.0

# Deep Learning & NLP
torch>=2.1.0
transformers>=4.36.0
datasets>=2.14.0
tokenizers>=0.15.0
sentence-transformers>=2.2.0
nltk>=3.8.1

# Data Processing & Storage
elasticsearch>=8.11.0
redis>=5.0.1
kafka-python>=2.0.2
pymongo>=4.6.0
dask>=2023.12.0
pyarrow>=14.0.1

# ML Metrics & Validation
scikit-learn>=1.3.0
optuna>=3.4.0
mlflow>=2.8.0
wandb>=0.15.0
deepchecks>=0.17.0
evidently>=0.4.0

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
aiohttp>=3.9.0

# Visualization & Reporting
plotly>=5.18.0
dash>=2.14.0
kaleido>=0.2.1
nbformat>=5.9.0
jupyter>=1.0.0
python-docx>=1.0.0
pdfkit>=1.0.0

# Monitoring & Logging
prometheus-client>=0.19.0
grafana-client>=2.3.0
python-json-logger>=2.0.0
opensearch-py>=2.3.0
elastic-apm>=6.18.0

# Testing & Quality
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
mypy>=1.7.0
pylint>=3.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
joblib>=1.3.0
psutil>=5.9.0
tenacity>=8.2.0